// src/app/models/book.ts
export class Book {
    title: String = 'title';
    description: String = 'description';
    thumbnail: String = 'https://external-preview.redd.it/m2WilH2KUvUQk98_CMFLUYVDLEKQb_UIEvdz2SHJilY.png?auto=webp&s=f0a7d47e3c88736bdd1b4e6c247a0a66cfd46a32';
    author: String = 'author';
    type: String = 'type';
}